+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+* More alerts endpoints
 * Deployments: purge cache endpoint
 
 ## 0.2.9